+++ Sources/Vapor/Client/Application+Clients.swift
@@ -4,17 +4,14 @@ extension Application {
     }
     
     public var client: Client {
-        clients.client
+        guard let makeClient = self.clients.storage.makeClient else {
+            fatalError("No client configured. Configure with app.clients.use(...)")
+        }
+        return makeClient(self)
     }
 
     public struct Clients {
         public struct Provider {
-            public static var http: Self {
-                .init {
-                    $0.clients.use { $0.clients.http }
-                }
-            }
-
             let run: (Application) -> ()
 
             public init(_ run: @escaping (Application) -> ()) {
@@ -31,17 +28,6 @@ extension Application {
             typealias Value = Storage
         }
 
-        public var http: AsyncHTTPClient {
-            return AsyncHTTPClient(eventLoop: self.application.eventLoopGroup.next(), application: self.application)
-        }
-        
-        public var client: Client {
-            guard let makeClient = self.storage.makeClient else {
-                fatalError("No client configured. Configure with app.clients.use(...)")
-            }
-            return makeClient(self.application)
-        }
-        
         public func initialize() {
             self.application.storage[Key.self] = .init()
             self.use(.http)
@@ -57,7 +43,7 @@ extension Application {
 
         public let application: Application
         
-        private var storage: Storage {
+        var storage: Storage {
             guard let storage = self.application.storage[Key.self] else {
                 fatalError("Clients not configured. Configure with app.clients.initialize()")
             }
