+++ Sources/Vapor/Client/Request+Client.swift
@@ -1,5 +1,5 @@
 extension Request {
     public var client: Client {
-        self.application.clients.client.for(self)
+        self.application.client.delegating(to: self.eventLoop)
     }
 }
