+++ Sources/Vapor/HTTP/Application+HTTP.swift
@@ -1,0 +1,9 @@
+extension Application {
+    public var http: HTTP {
+        .init(application: self)
+    }
+
+    public struct HTTP {
+        public let application: Application
+    }
+}
