+++ Sources/Vapor/Passwords/PlaintextPasswordVerifier.swift
@@ -1,54 +1,26 @@
-extension Application {
-    public var passwords: Passwords {
-        .init(application: self)
-    }
-
-    public struct Passwords {
-        public struct Provider {
-            let run: (Application) -> ()
-
-            public init(_ run: @escaping (Application) -> ()) {
-                self.run = run
+extension Application.Passwords.Provider {
+    public static var plaintext: Self {
+        .init {
+            $0.passwords.use { _ in
+                PlaintextPasswordVerifier()
             }
         }
-        
-        final class Storage {
-            var makeVerifier: ((Application) -> PasswordVerifier)?
-            init() { }
-        }
+    }
+}
 
-        struct Key: StorageKey {
-            typealias Value = Storage
-        }
-
-        let application: Application
-
-        public func use(_ provider: Provider) {
-            provider.run(self.application)
-        }
-
-        public func use(_ makeVerifier: @escaping (Application) -> (PasswordVerifier)) {
-            self.storage.makeVerifier = makeVerifier
-        }
-
-        func initialize() {
-            self.application.storage[Key.self] = .init()
-            // Default to BCrypt
-            self.use(.bcrypt)
-        }
-
-        var storage: Storage {
-            guard let storage = self.application.storage[Key.self] else {
-                fatalError("PasswordHashers not configured. Configure with app.passwordHashers.initialize()")
-            }
-            return storage
-        }
+struct PlaintextPasswordVerifier: PasswordVerifier {
+    func digest<Password>(_ password: Password) throws -> [UInt8]
+        where Password: DataProtocol
+    {
+        password.copyBytes()
     }
 
-    public var password: PasswordVerifier {
-        guard let makeVerifier = self.passwords.storage.makeVerifier else {
-            fatalError("No password verifier configured. Configure with app.passwords.use(...)")
-        }
-        return makeVerifier(self)
+    func verify<Password, Digest>(
+        _ password: Password,
+        created digest: Digest
+    ) throws -> Bool
+        where Password: DataProtocol, Digest: DataProtocol
+    {
+        password.copyBytes() == digest.copyBytes()
     }
 }
