+++ Sources/Vapor/View/Request+View.swift
@@ -1,0 +1,5 @@
+extension Request {
+    public var view: ViewRenderer {
+        self.application.view.for(self)
+    }
+}
