+++ Tests/VaporTests/ClientTests.swift
@@ -6,7 +6,7 @@ final class ClientTests: XCTestCase {
         let app = Application(.testing)
         defer { app.shutdown() }
 
-        app.clients.http.configuration.redirectConfiguration = .disallow
+        app.http.client.configuration.redirectConfiguration = .disallow
 
         app.get("redirect") {
             $0.redirect(to: "foo")
@@ -24,7 +24,7 @@ final class ClientTests: XCTestCase {
         let app = Application(.testing)
         defer { app.shutdown() }
 
-        app.clients.http.configuration.redirectConfiguration = .disallow
+        app.http.client.configuration.redirectConfiguration = .disallow
 
         app.get("redirect") {
             $0.redirect(to: "foo")
@@ -35,7 +35,7 @@ final class ClientTests: XCTestCase {
 
         _ = try app.client.get("http://localhost:8080/redirect").wait()
         
-        app.clients.http.configuration.redirectConfiguration = .follow(max: 1, allowCycles: false)
+        app.http.client.configuration.redirectConfiguration = .follow(max: 1, allowCycles: false)
         let res = try app.client.get("http://localhost:8080/redirect").wait()
         XCTAssertEqual(res.status, .seeOther)
     }
@@ -111,7 +111,7 @@ final class ClientTests: XCTestCase {
         Thread.async {
             startingPistol.leave()
             startingPistol.wait()
-            XCTAssert(type(of: app.clients.http) == AsyncHTTPClient.self)
+            XCTAssert(type(of: app.http.client.current) == HTTPClient.self)
             finishLine.leave()
         }
 
@@ -119,7 +119,7 @@ final class ClientTests: XCTestCase {
         Thread.async {
             startingPistol.leave()
             startingPistol.wait()
-            XCTAssert(type(of: app.clients.http) == AsyncHTTPClient.self)
+            XCTAssert(type(of: app.http.client.current) == HTTPClient.self)
             finishLine.leave()
         }
